Cura/requirements.txt
apns2==0.4.1
asn1crypto==0.24.0
astroid==2.1.0
bcrypt==3.1.6
certifi==2018.4.16
chardet==3.0.4
colorama==0.4.1
cryptography==2.4.2
cymysql==0.9.13
defusedxml==0.6.0
django-cymysql==2.1.0
django-excel==0.0.10
django-multiselectfield==0.1.8
django-paypal==1.0.0
django-push-notifications==1.6.0
django-rest-auth==0.9.3
django-simple-history==2.5.1
django-smtp-ssl==1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
et-xmlfile==1.0.1
h2==2.6.2
hpack==3.0.0
http-ece==1.0.5
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
isort==4.3.17
jdcal==1.4
lazy-object-proxy==1.3.1
lml==0.0.7
lxml==4.3.1
mccabe==0.6.1
openpyxl==2.6.1
passlib==1.7.1
paybook==2.5.6
paypal==1.2.5
Pillow==6.2.1
protobuf==3.6.1
py-vapid==1.4.0
pycparser==2.19
pyexcel==0.5.10
pyexcel-io==0.5.11
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
PyJWT==1.6.4
pyOpenSSL==19.0.0
python-decouple==3.1
pytz==2018.4
pywebpush==1.7.0
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
stripe==2.37.1
texttable==1.5.0
typed-ast==1.1.1
urllib3==1.24.1
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
